TerminalBrightness changed to double
This had to be done to work around a bug in stock KSP's code
that writes out config.xml files.  That code doesn't write out
anything to the file at all when you try to make it write a
single-precision float.

+++ src/kOS.Safe/Screen/ScreenBuffer.cs
@@ -29,7 +29,7 @@ namespace kOS.Safe.Screen
         public bool VisualBeep {get; set;}
         public int CharacterPixelWidth { get; set; }
         public int CharacterPixelHeight { get; set; }
-        public float Brightness { get; set; }
+        public double Brightness { get; set; }
 
         
         public int ColumnCount { get; private set; }
